+++ CHANGELOG.md
@@ -45,6 +45,8 @@ Highlights are marked with a pancake 🥞
 - sync: Method for streaming log ranges from the store [#1393](https://github.com/p2panda/p2panda/pull/1393)
 - sync: Method to resolve log heights from topic [#1405](https://github.com/p2panda/p2panda/pull/1405)
 - sync: Re-export -stream ingest_operation in `api` module [#1406](https://github.com/p2panda/p2panda/pull/1406)
+- stream: Forward events of already processed operations in spaces processor [#1430](https://github.com/p2panda/p2panda/pull/1430)
+- p2panda: Support optionally adding spaces events when constructing processor Event [#1430](https://github.com/p2panda/p2panda/pull/1430)
 
 ### Changed